% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function2.R
\name{my_function2}
\alias{my_function2}
\title{Name of my cats}
\usage{
my_function2(aname, bname)
}
\arguments{
\item{aname}{name of first cat}

\item{bname}{name of second cat}
}
\value{
a sentence with the name of both of my cats
}
\description{
Name of my cats
}
